Move text_split
